pyproject.toml
[tool.poetry]
name = "tcc"
version = "0.1.0"
description = ""
authors = ["VictorAugustoBarros <victor.augustobarros@gmail.com>"]
packages = [
    { include = "connections", from = "app" },
    { include = "models", from = "app" },
    { include = "routes", from = "app" },
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"
pyArango = "^2.0.1"
aioodbc = "^0.3.3"
numpy = "^1.24.2"
pyjwt = "^2.7.0"
pandas = "^2.0.1"
ahpy = "^2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api = 'app.api:api.run'